Moved experience to its own page
