+++ lib/src/localization/app_en.arb
@@ -7,5 +7,6 @@
   "profilePageTitle": "Profile Page",
   "experienceSummaryText": "Flutter Developer 💙. 8 years with JavaScript. 3 Years with Flutter :)",
   "viewMyArticles": "My articles",
-  "viewShortPosts": "Short posts"
+  "viewShortPosts": "Short posts",
+  "viewExperiencePage": "Professional Experience"
 }
