+++ lib/src/localization/app_uk.arb
@@ -7,5 +7,6 @@
   "profilePageTitle": "Профіль",
   "experienceSummaryText": "Flutter Розробник 💙. 8 років JavaScript. 3 роки з Flutter :)",
   "viewMyArticles": "Мої статті",
-  "viewShortPosts": "Короткі дописи"
+  "viewShortPosts": "Короткі дописи",
+  "viewExperiencePage": "Професійний досвід"
 }
